Added Desired Functions pluss ekstra
Added the desired functions plus some extra function's that could be used if there is enough time.

The ArduinoSumoCode folder the primary folder that people are going to use.
The ArduinoSumoCode is a little more advanced containing If statements if we need extra stuff do bc there is enough time